backend/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.6,<5.0
python-dotenv>=1.0,<2.0
google-genai>=1.0,<2.0
openai>=1.0,<2.0
pytest>=8.0,<9.0
pytest-django>=4.9,<5.0
